+++ core/interaction-api/src/main/java/ru/practicum/core/interactionapi/dto/EventShortDto.java
@@ -12,10 +12,10 @@ public class EventShortDto {
     private Long id;
     private String annotation;
     private CategoryDto category;
-    private Integer confirmedRequests;
+    private Long confirmedRequests;
     private String eventDate;
     private UserShortDto initiator;
     private Boolean paid;
     private String title;
-    private Integer views;
+    private Long views;
 }
